TFDS-Create and share a Jupyter Notebook
This is a Notebook created as a part of the Tools for data science course as the Final Assignment in Moduel 6. The objective for this assignment is to create a Jupyter Notebook, insert code- and markdows cells according to instructions and Share this trough GitHub. The purpose of sharing this is to get peer review for grading of the assignment.